Final type and presenter lookup..
constants, const_defined? and defined? all return false for reloaded constants. This caused a problem with an attribute presenter in the host app, since it was in the autoload paths in app/presenters. Safe-calling the constants with defined must skip the rails class loader, so checking them with a rescue seems to be the only way to reliably determine what exists