Complete DSL refactor
Controller can now configure modules using a block, which class_execs a dynamic subclass of the module's builder, or using their own subclass.

Configuration creates builders, which can be configured multiple times then built at the end of each configure cycle

Methods can be defined on the builder to encapsulate configuration. This is called by an instance, passing the builder it is intended to configure.